Crate aws_mpl_primitives

aws-mpl-primitives

The aws-mpl-primitives crate wraps the aws-lc-rs crate, providing a more convenient interface.

This crate is intended for internal use by aws-esdk, aws-db-esdk and aws-mpl-rs.

Features

fips

enable the fips feature in the aws-lc-rs crate

track

Enable the #[global_allocator] for use with memory_tracker::ResourceTracker

Modules

format

format numbers into human readable abbreviations

memory_tracker

Track memory allocations. Use track feature to turn on the global allocator tracking.

Structs

DigestContext

Gathers Digest Data

Error

Error type for this crate

Prk

The intermediate value, generated by extract, used by expand

Enums

AesGcm

AES GCM Algorithm. Default is Aes256Gcm

DigestAlg

Digest Algorithm. Default is Sha512

EcdhCurveSpec

EcdsaSignatureAlgorithm

Which ECDSA Algorithm to use

Functions

aes_decrypt

Decrypt message with AES GCM

aes_encrypt

Encrypt message with AES GCM

constant_time_equal

Compare in constant time

ctr_stream

AES KDF CTR

digest

Return a digest of the data using the selected algorithm

ecdsa_sign

Sign the message with the key using the given algorithm

ecdsa_sign_digest

Sign the message with the key using the given algorithm

ecdsa_verify

Verify the signature of the message using the key and algorithm

ecdsa_verify_context

Verify the signature of the message using the key and algorithm

generate_random_bytes

Fill bytes with random data.

hkdf

Calculate HKDF

hkdf_expand

HKDF Expand

hkdf_extract

HKDF Extract

hkdf_no_salt

Calculate HKDF, with default salt

hmac

Calculate HMAC